Linked_List_Linear_Search-1_27947
A linked list is used to store Roll Numbers of N students. Priyanka want to check that whether data is stored in palindrome format or not. Write a program to help the priyanka which will provide result in 'Yes' or 'No' depending on the data stored in list is in palindrome format.