array_bubblesort-2_26699
In a sports event, Lindsey, a famous athlete, has decided to meet her fans and shake hands with them. The order in which she will meet them is the fan degree. i.e. the fan with maximum degree will be considered first. Now, the fans are required to take positions according to the degree value. While they are not allowed to swap positions with any other person, but with the person at their adjacent places. Write a program to implement the same.