delete_node_linked_list_1_26199
Write a program to delete the node from given position in the linked list. The list may be empty after you delete the node. In that case, print "SLL is Empty". If position of linked list is not valid then print a message of "Invalid Position"


Use the following structure to implement it.
struct node SLLNode
{
int data;
SLLNode* next;
};

Input Format

The first line of input contains the number of elements in the linked list.
The second line of input contains an space separeted node data values.
The Third line contains the position of the node(position number starts from 0) to delete.

Constraints

0<=n<=1000, where n is number of elements in the linked list.
the data part of node also must be in between 0 to 1000.

Output Format

It will display the singly linked list after deleting the node from the singly linked list.
The list may be empty after you delete the node. In that case, print "SLL is Empty". If position of linked list is not valid then print a message of "Invalid Position"


Sample Input 0

8
2 6 5 1 7 4 15 9
3
Sample Output 0

2 6 5 7 4 15 9
Explanation 0

The original list is 2->6->5->1->7->4->15->9.
After deleting the node at position 3, the list is 2->6->5->7->4->15->9.